requirements.txt
fastapi
uvicorn

langgraph
langchain
langchain-google-genai
google-generativeai

httpx
python-dotenv

prometheus-client
pydantic
pydantic-settings

pytest
pytest-asyncio